done for now

+++ src/screens/SignUp.tsx
@@ -3,19 +3,17 @@ import { View, KeyboardAvoidingView, Animated, Text } from 'react-native';
 import AsyncStorage from '@react-native-community/async-storage';
 import { Input, Button } from 'react-native-elements';
 import Icon from 'react-native-vector-icons/FontAwesome';
-import { useMutation } from '@apollo/react-hooks';
 import { MainContext } from '../store';
 import styles from '../styles';
 import { registerForPushNotificationsAsync } from '../utils/helpers';
-import {createUser} from '../api/api'
+import { createUser } from '../api/api';
 import jwt_decode from 'jwt-decode';
 
-
-const SignUp = (props) => {
+const SignUp = ({ navigation }) => {
   const [name, setName] = useState('');
   const [password, setPassword] = useState('');
   const [confirmPassword, setConfirmPassword] = useState('');
-  const [loading, setLoading] = useState(false)
+  const [loading, setLoading] = useState(false);
   const [email, setEmail] = useState('');
   const [error, setError] = useState('');
   const globalState = useContext(MainContext);
@@ -30,55 +28,56 @@ const SignUp = (props) => {
     name: string;
     iat: number;
     exp: number;
-    push_token: string
+    push_token: string;
   }
-
 
   const onSubmit = async () => {
     const push_token = await registerForPushNotificationsAsync();
+    if (email && name && password && confirmPassword) {
+      setDisableButton(true);
 
-
-      const data = await createUser(email,
-        name,
-        password,
-        push_token);
-      if(data.error){
-        setError(data.error.message)
-      }else{
+      const data = await createUser(email, name, password, push_token);
+      if (data.error) {
+        setError(data.error.message);
+        setDisableButton(false);
+      } else {
         setDisableButton(false);
         setError('');
+
+        const decoded: Decoded = jwt_decode(data.token);
+
+        console.log('data.token', data.token);
+        console.log('decoded.push_token', decoded.push_token);
+        console.log('decoded.email', decoded.email);
+        console.log('decoded.user_id', decoded.user_id);
+        console.log('decoded.name', decoded.name);
+
+        try {
+          await AsyncStorage.setItem('AUTH_TOKEN', data.token);
+          await AsyncStorage.setItem('PUSH_TOKEN', push_token);
+          await AsyncStorage.setItem('EMAIL', decoded.email);
+          await AsyncStorage.setItem('USER_ID', decoded.user_id);
+          await AsyncStorage.setItem('NAME', decoded.name);
+        } catch (e) {
+          console.log(e);
+        }
+
+        dispatch({
+          type: 'SET_USER',
+          payload: {
+            user_id: decoded.user_id,
+            token: data.token,
+            push_token: decoded.push_token,
+          },
+        });
+        setEmail('');
+        setPassword('');
+        setName('');
+        setConfirmPassword('');
+        if (!error) {
+          navigation.navigate('NavDrawer', { screen: 'Map' });
+        }
       }
-
-
-    const decoded: Decoded = jwt_decode(data.token);
-
-
-   
-
-    try {
-      await AsyncStorage.setItem("AUTH_TOKEN", data!.token);
-      await AsyncStorage.setItem("PUSH_TOKEN", decoded.push_token);
-      await AsyncStorage.setItem("EMAIL", decoded.email);
-      await AsyncStorage.setItem("USER_ID", decoded.user_id);
-      await AsyncStorage.setItem("NAME", decoded.name);
-    } catch (e) {
-      return e;
-    }
-
-    dispatch({
-      type: "SET_USER",
-      payload: {
-        user_id: decoded.user_id,
-        token: token,
-        push_token: decoded.push_token
-      },
-    });
-    setEmail("");
-    setPassword("");
-    setName("");
-    setConfirmPassword("");
-    if (!error) {
-      props.navigation.navigate("NavDrawer", { screen: "Map" });
     }
   };
 
@@ -151,11 +150,10 @@ const SignUp = (props) => {
         style={styles.inputContainerStyle}
       />
       <Button
-        icon={<Icon name="arrow-right" size={15} color="white" />}
         title="Submit"
         buttonStyle={styles.signupSubmit}
         onPress={() => onSubmit()}
-        disabled={loading}
+        disabled={disableButton}
       />
     </KeyboardAvoidingView>
   );
